Fix GLenum arithmetic (GL_COLOR_ATTACHMENT0+i, GL_DRAW_BUFFER0+i), GLbitfield bitmask types
Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/9f0a555b-ffc1-4c26-b332-6bd707db7944

+++ modules/basegl/src/processors/volumeprocessing/volumeregionshrink.cpp
@@ -276,7 +276,8 @@ void VolumeRegionShrink::process() {
     size_t dst = 0;
     for (int i = 1; i < iterations_; ++i) {
         std::swap(src, dst);
-        glDrawBuffer(static_cast<GLenum>(GL_COLOR_ATTACHMENT0 + dst));
+        glDrawBuffer(static_cast<GLenum>(static_cast<GLuint>(GL_COLOR_ATTACHMENT0) +
+                                       static_cast<GLuint>(dst)));
         TextureUnitContainer cont;
         utilgl::bindAndSetUniforms(shader_, cont, *out_[src], "volume");
         utilgl::setShaderUniforms(shader_, *out_[dst], "dstParameters");
